Accessing Data in Code Segments
In some instances it may be desirable to access data structes that are
contained in a code segment. The following methods of accessing data in code
segments are possible:

* Load a data-segment register with a segment selector for a nonconforming,
  readable, code segment.

* Load a data-segment register with a segment selector for a conforming,
  readable, code segment.

* User a code-segment override prefix(CS) to read a readable, code segment
  whose selector is already loaded in the CS register.

The same rules for accessing data segments apply to method 1. Method 2 is
always valid because the privilege level of a conforming code segment is
effectively the same as the CPL, regardless of its DPL. Method 3 is always
valid because the DPL of the code segment selected by the CS register is the
same as the CPL.