Accessing Conforming Code Segment
When accessing conforming code segments, the CPL of the
calling procedure may be numerically equal to or greater
than (less privileged) the DPL of the destination code
segment; the processor generates a general-protection
exception (#GP) only if the CPL is less than the DPL.
(The segment selector RPL for the destination code segment
is not checked if the segment is a conforming code segment.)

In the example in Figure, code segment D is a conforming
code segment.Therefore, calling procedures in both code
segment A and B can access code segment D (using eigher
segment selector D1 or D2, respectively), because they both
have CPLs that are greater than or equal to the DPL of the
conforming code segment. For conforming code segment, the
DPL represents the numerically lowest privilege level that
calling procedure may be at to successful make a call to the
code segment.