Accessing a Code Segment Though a Call Gate
To access a call gate, a far pointer to the gate is provided
as a target operand in a CALL or JMP instruction. The segment
selector from this pointer identifies the call gate (See
Figure); the offset from the pointer is required, but not used
or checked by the processor. (The offset can be set to any value.)

When the processor has accessed the call gate, it uses the
segment selector from the call gate to locate the segment
descriptor for the destination code segment. (This segment
descriptor can be in the GDT or the LDT.) It then combines
the base address from the code-segment descriptor with the
offset from the call gate to form the linear address of the
procedure entry point in the code segment.