Privilege check accessing CALL GATES
The DPL field of the call-gate descriptor specifies the
numbercally highest privilege from which a calling procedure
can access the call gate; that is, to access a call gate, the
CPL of a calling procedure must be equal to or less than the
DPL of the call gate. For example, in Figure, call gate A has
a DPL of 3. So calling procedure at all CPLs (0 through 3) can
access this call gate, which includes calling procedure in code
segments A, B, and C. Call gate B has DPL of 2, so only calling
procedures at a CPL or 0,1, or 2 can access call gate B, which
includes calling procedures in code segments B and C. The dotted
line shows that a calling procedure in code segment A cannot
access call gate B.

The RPL of the segment selector to a call gate must satisfy
the same test as the CPL of the calling procedure; that is,
the RPL must be less than or equal to the DPL of the call gate.
In the example in Figure, a calling procedure in code segment C
can access call gate B using gate selector B2 or B1, but it
could not use gate selector B3 to access call gate B.

If the privilege checks between the calling procedure and call
gate are successful, the processor then checks the DPL of the
code-segment descriptor against the CPL of the calling procedure.
Here, the privilege check rules vary between CALL and JMP
instructions. Only CALL instruction can use call gates to tranfer
program control to more privilege (numerically lower privilege
level) nonconforming code segment; that is, to nonconforming code
segmetn with a DPL less than the CPL. A JMP instrcution can use a
call gate only to transfer program control to a nonconforming code
segment with a DPL equal to the CPL. CALL and JMP instruction can
both transfer program control to a more privileged conforming code
segment; that is, to a conforming code segment with a DPL less
than or equal to the CPL.